initial stab at enabling new registrants for free trial
