+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/TrialServiceSpec.scala
@@ -7,29 +7,32 @@ import org.broadinstitute.dsde.firecloud.model.{Trial, WorkbenchUserInfo}
 import org.broadinstitute.dsde.rawls.model.RawlsBillingProjectName
 import org.scalatest.BeforeAndAfterEach
 
+import scala.concurrent.ExecutionContext
+
 class TrialServiceSpec extends BaseServiceSpec with BeforeAndAfterEach with TrialServiceSupport {
 
-  val trialDAO = new TrialServiceSpecTrialDao
+  override val trialDao = new TrialServiceSpecTrialDao
+  implicit protected val executionContext: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
 
   val userInfo = WorkbenchUserInfo("foo", "bar")
 
   override protected def beforeEach() = {
-    trialDAO.claimCalled = false
+    trialDao.claimCalled = false
   }
 
   "TrialService.buildEnableUserStatus" - {
     "should claim a ProjectRecord when starting from trial status with None state" in {
       buildEnableUserStatus(userInfo, UserTrialStatus("userid", None, true, 0,0,0,0,None))
-      assert(trialDAO.claimCalled)
+      assert(trialDao.claimCalled)
     }
     "should claim a ProjectRecord when starting from Disabled" in {
       buildEnableUserStatus(userInfo, UserTrialStatus("userid", Some(Disabled), true, 0,0,0,0,None))
-      assert(trialDAO.claimCalled)
+      assert(trialDao.claimCalled)
     }
     Seq(Enabled, Enrolled, Terminated) foreach { state =>
       s"should not claim a ProjectRecord when starting from $state" in {
         buildEnableUserStatus(userInfo, UserTrialStatus("userid", Some(state), true, 0,0,0,0,None))
-        assert(!trialDAO.claimCalled)
+        assert(!trialDao.claimCalled)
       }
     }
   }
